sign in functionality complete
