+++ back-end/mock-data/currentUser.json
@@ -1,0 +1,12 @@
+[
+    {
+      "id": "user_123",
+      "username": "john_doe",
+      "password": "password123",
+      "email": "john@example.com",
+      "name": "John Doe",
+      "profileAvatar": "😎",
+      "bio": "Travel enthusiast and food lover.",
+      "trips": ["trip_456", "trip_789", "trip_101"]
+    }
+]
