+++ front-end/src/App.js
@@ -14,40 +14,8 @@ import LogIn from './pages/LogIn';
 import SignUp from './pages/SignUp';
 
 const App = () => {
-  const [isLoggedIn, setIsLoggedIn] = useState(true); // Default to "logged in" for mock
+  const [isLoggedIn, setIsLoggedIn] = useState(false); // Default to "logged in" for mock
   const [user, setUser] = useState(null);
-  const [users, setUsers] = useState([]); // State to store all users
-
-
-  useEffect(() => {
-    // Fetch user data when the component mounts
-    fetchUserData();
-
-    // Fetch all users when the app launches
-    fetchAllUsers();
-  }, []);
-
-  const fetchAllUsers = async () => {
-    try {
-      const response = await fetch('/users');
-      const data = await response.json();
-      setUsers(data); // Set all users in the state
-      console.log('All users:', data);
-    } catch (error) {
-      console.error('Error fetching users:', error);
-    }
-  };
-
-  const fetchUserData = async () => {
-    try {
-      const response = await fetch('/users/64b1c7c8f2a5b9a2d5c8f001');
-      const data = await response.json();
-      setUser({ name: data.name, profileAvatar: data.profileAvatar });
-      console.log('Current user:', data);
-    } catch (error) {
-      console.error('Error fetching user data:', error);
-    }
-  };
 
   const handleSignOut = () => {
     setIsLoggedIn(false);
@@ -73,7 +41,9 @@ const App = () => {
           <Route path="/profile" element={<ProfilesPage setUser={setUser} />} />
           <Route path="/create-trip/:userId" element={<AddTrip />} />
           <Route path="/join-trip" element={<JoinTrip />} />
-          <Route path="/log-in" element={<LogIn />} />
+          {/* <Route path="/log-in" element={<LogIn />} /> */}
+          <Route path="/log-in" element={<LogIn user={user} setUser={setUser} setIsLoggedIn={setIsLoggedIn} />} />
+
           <Route path="/sign-up" element={<SignUp />} />
         </Routes>
       </main>
